Make buckets a separate type
